# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Payment gateway aggregator — VA, QRIS, and disbursement via Midtrans, Xendit, DOKU, iPaymu.\n\n## Authentication\nMerchant endpoints (/v1/*): `X-API-Key` header with your API key (`wpay_live_*` or `wpay_test_*`).\nAdmin endpoints (/admin/*): `Authorization: Bearer \u003caccess_token\u003e` from POST /admin/login.\n\n## Idempotency\nPayment and disbursement creation support `Idempotency-Key` header. Use unique key per request to safely retry.\n\n## Webhook\nWanpey signs outbound webhooks with `X-Wanpey-Signature` (HMAC-SHA256). Verify using your webhook secret.",
        "title": "Wanpey Payment Gateway API",
        "contact": {
            "name": "Wanpey Support",
            "email": "support@wanpey.id"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/disbursements": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns paginated list of disbursements for the authenticated merchant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disbursements"
                ],
                "summary": "List disbursements",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status: pending, processing, completed, failed",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (RFC3339)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.ListResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/wanpey_core_internal_usecase.DisbursementOutput"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Sends funds from merchant balance to a registered verified bank account. The bank account must be verified by admin before use. Daily cashout limit applies if configured.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disbursements"
                ],
                "summary": "Create disbursement (cash out)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique key to prevent duplicate disbursements on retry",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Disbursement request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_internal_usecase.DisburseInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wanpey_core_internal_usecase.DisbursementOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Insufficient balance",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Bank account not verified or daily limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/disbursements/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns full disbursement detail including status, bank account info, and fee breakdown.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disbursements"
                ],
                "summary": "Get disbursement detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Disbursement ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wanpey_core_internal_usecase.DisbursementOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "wanpey_core_internal_domain_entity.BankCode": {
            "type": "string",
            "enum": [
                "BCA",
                "BNI",
                "BRI",
                "BSI",
                "MANDIRI",
                "PERMATA",
                "CIMB"
            ],
            "x-enum-varnames": [
                "BankBCA",
                "BankBNI",
                "BankBRI",
                "BankBSI",
                "BankMandiri",
                "BankPermata",
                "BankCIMB"
            ]
        },
        "wanpey_core_internal_domain_entity.Currency": {
            "type": "string",
            "enum": [
                "IDR"
            ],
            "x-enum-varnames": [
                "CurrencyIDR"
            ]
        },
        "wanpey_core_internal_domain_entity.DisbursementStatus": {
            "type": "string",
            "enum": [
                "pending",
                "processing",
                "completed",
                "failed"
            ],
            "x-enum-varnames": [
                "DisbursementStatusPending",
                "DisbursementStatusProcessing",
                "DisbursementStatusCompleted",
                "DisbursementStatusFailed"
            ]
        },
...
        "wanpey_core_internal_domain_entity.Provider": {
            "type": "string",
            "enum": [
                "midtrans",
                "xendit",
                "doku",
                "ipaymu"
            ],
            "x-enum-varnames": [
                "ProviderMidtrans",
                "ProviderXendit",
                "ProviderDoku",
                "ProviderIPaymu"
            ]
        },
...
        "wanpey_core_internal_usecase.DisburseInput": {
            "type": "object",
            "required": [
                "amount",
                "bank_account_id",
                "currency",
                "provider"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "bank_account_id": {
                    "type": "string"
                },
                "currency": {
                    "enum": [
                        "IDR"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/wanpey_core_internal_domain_entity.Currency"
                        }
                    ]
                },
                "description": {
                    "type": "string",
                    "maxLength": 255
                },
                "provider": {
                    "enum": [
                        "xendit",
                        "doku"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/wanpey_core_internal_domain_entity.Provider"
                        }
                    ]
                }
            }
        },
        "wanpey_core_internal_usecase.DisbursementOutput": {
            "type": "object",
            "properties": {
                "account_name": {
                    "type": "string"
                },
                "account_number": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "bank_code": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.BankCode"
                },
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.Currency"
                },
                "description": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "failure_reason": {
                    "type": "string"
                },
                "fee_amount": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "provider": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.Provider"
                },
                "status": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.DisbursementStatus"
                }
            }
        },
...
        "wanpey_core_pkg_response.ErrorBody": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wanpey_core_pkg_response.FieldDetail"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "wanpey_core_pkg_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.ErrorBody"
                },
                "meta": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "wanpey_core_pkg_response.FieldDetail": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "wanpey_core_pkg_response.ListResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "wanpey_core_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Pagination"
                },
                "timestamp": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
        "wanpey_core_pkg_response.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "wanpey_core_pkg_response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin JWT — Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ApiKeyAuth": {
            "description": "Merchant API key",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}